requirements.txt
protobuf~=5.29.5
Flask~=3.1.2
google-generativeai